requirements.txt
transformers==4.20.1
kiwipiepy==0.13.1
pymysql==1.0.2
google-cloud-translate
flask
nltk